Updated Version of Example File - More Indepth
I am uploading a new version of our example.ipynb file. This file now contains more indepth explanations of our functions and use cases, they contain real world examples and context as to how to use the functions, as well as explanations of the input and output values. Will need someone else to run it though and produce the outputs.