Setup GitHub actions workflow for continuous deploymen
Hello guys, to setup GitHub Action workflows so that the package is built and published to PyPI when pull requests are accepted to the main branch, I create the template the ci-cd.yml file from the py-pkgs-cookiecutter repository as demonstrated in the lecture 7. Please take a look to see if it is ok to be merged. Thanks!